+++ changelog.md
@@ -6,6 +6,7 @@
 - parsing: fix %_Q %-Q %_q %-q
 - formatting: fix %3ES %3Es
 - change internal members of ParseTime to allow newtype-deriving
+- new functions (aliases) pastMidnight & sinceMidnight
 
 ## [1.9.3]
 - documentation fixes